File Uploads and Intelligent Processing/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.1
openai==1.3.0
supabase==2.0.2
numpy==1.24.3
python-dotenv==1.0.0
werkzeug==2.3.7
